+++ src/backend/core/api/resource_server/scim/__init__.py
@@ -1,0 +1,1 @@
+"""SCIM compliant API for resource server"""